+++ src/modules/mavlink/mavlink_messages.h
@@ -49,9 +49,9 @@ class StreamListItem
 public:
 	MavlinkStream *(*new_instance)(Mavlink *mavlink);
 	const char *(*get_name)();
-	uint8_t (*get_id)();
+	uint16_t (*get_id)();
 
-	StreamListItem(MavlinkStream * (*inst)(Mavlink *mavlink), const char *(*name)(), uint8_t (*id)()) :
+	StreamListItem(MavlinkStream * (*inst)(Mavlink *mavlink), const char *(*name)(), uint16_t (*id)()) :
 		new_instance(inst),
 		get_name(name),
 		get_id(id) {};
